# api.yaml
openapi: 3.0.1
info:
  title: SpeedQuiz API
  description: API for managing quiz decks and cards
  version: 2.1.0

tags:
  - name: deck
    description: Deck listing, loading, browsing, creation, and card state management

paths:
  /quizApi/decks:
    get:
      tags:
        - deck
      summary: List all available decks
      operationId: listDecks
      parameters:
        - name: ownerId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: wkClaimedName
          in: query
          required: false
          schema:
            type: string
        - name: wkTokenHash
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A list of available decks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeckInfo'

    post:
      tags:
        - deck
      summary: Create a new deck
      operationId: createDeck
      parameters:
        - name: ownerId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: deckName
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeckContent'
      responses:
        '201':
          description: Deck created successfully

  /quizApi/decks/{deckId}:
    get:
      tags:
        - deck
      summary: Load a full deck
      operationId: loadDeck
      parameters:
        - name: deckId
          in: path
          required: true
          schema:
            type: string
        - name: ownerId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: wkClaimedName
          in: query
          required: false
          schema:
            type: string
        - name: wkApiToken
          in: query
          required: false
          schema:
            type: string
        - name: wkTokenHash
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Full deck content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckContent'

  /quizApi/decks/{deckId}/page:
    get:
      tags:
        - deck
      summary: Browse a deck with pagination and filtering
      operationId: browseDeck
      parameters:
        - name: deckId
          in: path
          required: true
          schema:
            type: string
        - name: ownerId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: filter
          in: query
          required: false
          schema:
            type: string
        - name: wkClaimedName
          in: query
          required: false
          schema:
            type: string
        - name: wkApiToken
          in: query
          required: false
          schema:
            type: string
        - name: wkTokenHash
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A page of cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckPage'

  /quizApi/decks/{deckId}/state:
    get:
      tags:
        - deck
      summary: Get card states for a deck
      operationId: getCardStates
      parameters:
        - name: deckId
          in: path
          required: true
          schema:
            type: string
        - name: ownerId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Card states
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeckCardState'

    post:
      tags:
        - deck
      summary: Update card states
      operationId: updateCardStates
      parameters:
        - name: deckId
          in: path
          required: true
          schema:
            type: string
        - name: ownerId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/DeckCardState'
      responses:
        '200':
          description: States updated

components:
  schemas:
    DeckInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        attribution:
          type: string
      required:
        - id
        - name

    DeckContent:
      type: object
      properties:
        properties:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/PropertyType'
        cards:
          type: array
          items:
            type: object
            additionalProperties:
              type: string
      required:
        - properties
        - cards

    DeckPage:
      type: object
      properties:
        cards:
          type: array
          items:
            type: object
            additionalProperties:
              type: string
        totalCards:
          type: integer
        offset:
          type: integer
        limit:
          type: integer

    DeckCardState:
      type: object
      properties:
        deckId:
          type: string
        cardId:
          type: string
        state:
          type: string

    PropertyType:
      type: string
      enum:
        - question
        - answer
        - hint
        - image
        - svg
        - audio
        - hiragana
        - answer_list
        - info
        - unknown